+++ crates/compiler/src/compiler.rs
@@ -1,4 +1,4 @@
-use crate::compiler::compilation_job::CompilationJob;
+pub use crate::compiler::compilation_job::*;
 use crate::output::*;
 use crate::parser::generated::yarnspinnerlexer::YarnSpinnerLexer;
 use crate::parser::generated::yarnspinnerparser::YarnSpinnerParser;
@@ -7,6 +7,7 @@ use antlr_rust::InputStream;
 
 mod compilation_job;
 
+/// Compile Yarn code, as specified by a compilation job.
 pub fn compile(_compilation_job: CompilationJob) -> CompilationResult {
     let lexer = YarnSpinnerLexer::new(InputStream::new(
         "# hello
@@ -23,5 +24,14 @@ Wow!
     let _dialogue = parser
         .dialogue()
         .expect("This error should be handled by the error listener and go into the diagnostics.");
-    todo!()
+
+    CompilationResult {
+        program: None,
+        string_table: Default::default(),
+        declarations: None,
+        contains_implicit_string_tags: false,
+        file_tags: Default::default(),
+        diagnostics: vec![],
+        debug_info: Default::default(),
+    }
 }
